Bug Corigé

+++ olympics/__main__.py
@@ -3,7 +3,7 @@
 import argparse
 
 from . import cli
-
+#Bug Corigé
 parser = argparse.ArgumentParser(
     prog='olympics',
     description='Display various information about Olympics results',
@@ -16,7 +16,7 @@ parser.add_argument(
 parser.add_argument(
     '--top',
     help='number of top elements to display',
-    type=float,
+    type=int,
     default=10,
 )
